Fixed unknown OS fallback error.
This commit fixes the graceful fallback when the OS can't be detected.
Before this fix the library would not compile.

The existing construct was changed to use the `cfg` macro to make the
code more readable, easier to maintain and easier to specify an `else`
clause (instead of having to manually specify an opposite condition of
all of the above branches).